doc: add online demo link for Sample 2 in README
Adds a link to a hosted build of Sample 2, allowing users to
explore the connection flow without importing the package. The
addition is made to both the English and Traditional Chinese
documentation files. This change is intended to enhance user
experience by providing an easy way to try out the functionalities
without the need for setup.